refactor: 뷰모델 스코프 모듈 분리
