+++ android/festago/app/src/main/java/com/festago/festago/data/di/ViewModelScopeModule.kt
@@ -1,0 +1,30 @@
+package com.festago.festago.data.di
+
+import com.festago.festago.data.repository.ReservationTicketDefaultRepository
+import com.festago.festago.data.repository.StudentVerificationDefaultRepository
+import com.festago.festago.data.repository.UserDefaultRepository
+import com.festago.festago.repository.ReservationTicketRepository
+import com.festago.festago.repository.StudentVerificationRepository
+import com.festago.festago.repository.UserRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ViewModelComponent
+import dagger.hilt.android.scopes.ViewModelScoped
+
+@InstallIn(ViewModelComponent::class)
+@Module
+interface ViewModelScopeModule {
+
+    @Binds
+    @ViewModelScoped
+    fun bindsReservationTicketDefaultRepository(reservationTicketRepository: ReservationTicketDefaultRepository): ReservationTicketRepository
+
+    @Binds
+    @ViewModelScoped
+    fun bindsStudentVerificationDefaultRepository(studentVerificationRepository: StudentVerificationDefaultRepository): StudentVerificationRepository
+
+    @Binds
+    @ViewModelScoped
+    fun binsUserDefaultRepository(userRepository: UserDefaultRepository): UserRepository
+}
